EAGLE-167 Remove deprecated hadoop_jmx_kafka.py and metric_extensions.py
